✨ feat : 팀 참가, 수정 페이지 구현 및 페이지 경로 수정
