+++ app/(team)/(team-potal)/create-team/layout.tsx
@@ -1,0 +1,6 @@
+export default function CreateTeamLayout({ children }: Readonly<React.PropsWithChildren>) {
+	return children;
+}
+
+/** @type {import("next").Metadata} */
+export const metadata = { title: "팀 생성하기" };
