+++ app/(team)/(team-potal)/join-team/layout.tsx
@@ -1,0 +1,6 @@
+export default function JoinTeamLayout({ children }: Readonly<React.PropsWithChildren>) {
+	return children;
+}
+
+/** @type {import("next").Metadata} */
+export const metadata = { title: "팀 참가하기" };
