+++ app/(team)/[id]/edit/layout.tsx
@@ -1,0 +1,6 @@
+export default function EditTeamLayout({ children }: Readonly<React.PropsWithChildren>) {
+	return children;
+}
+
+/** @type {import("next").Metadata} */
+export const metadata = { title: "팀 수정" };
