--- app/(team)/create-team/page.tsx
@@ -1,6 +1,0 @@
-export default function Page() {
-	return <div>WIP</div>;
-}
-
-/** @type {import("next").Metadata} */
-export const metadata = { title: "팀 생성하기" };
